Create code
In this code, we define a Bicycle class with a private member variable gear. The Bicycle class has two member functions: changeGear to set the gear to a new value and printGear to print the current gear.

In the main function, we create an instance of the Bicycle class called bike. We start by printing the initial gear, then use the changeGear function to switch gears, and finally, print the updated gear after each change.

When you run this code, you will see the output showing the current gear at each step of the gear change process. Feel free to modify the code to add more functionality or customize it to suit your needs.